requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
pytesseract
pillow
python-dotenv
mysql-connector-python
